requirements.txt
click==6.6
enum34==1.1.6
Flask==0.11.1
google-api-python-client==1.6.1
google-auth==0.5.0
google-auth-httplib2==0.0.2
google-cloud-core==0.22.1
google-cloud-vision==0.22.0
googleapis-common-protos==1.5.0
gunicorn==19.6.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2client==4.0.0
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
requests==2.10.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
Werkzeug==0.11.10
